requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
motor==3.3.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pydantic==2.5.0
pydantic-settings==2.1.0
spacy==3.7.2
transformers==4.36.2
torch==2.2.0
scikit-learn==1.3.2
sentence-transformers==2.2.2
PyPDF2==3.0.1
python-docx==1.1.0
pandas==2.1.4
numpy==1.24.4
python-dotenv==1.0.0
aiofiles==23.2.1
Pillow==10.1.0

# Advanced ML and AI dependencies
scikit-learn==1.3.2
xgboost==2.0.2
lightgbm==4.1.0
optuna==3.4.0

# LLM Integration
openai==1.3.7
anthropic==0.7.8
langchain==0.0.350
transformers==4.36.2
accelerate==0.25.0

# OCR and Multi-modal
pytesseract==0.3.10
opencv-python==4.8.1.78
pdf2image==1.16.3
easyocr==1.7.0

# Multilingual Support
googletrans==4.0.0rc1
langdetect==1.0.9
polyglot==16.7.4

# Knowledge Graph and Ontology
rdflib==7.0.0
owlready2==0.45
networkx==3.2.1
sparqlwrapper==2.0.0

# Bias Detection and Fairness
fairlearn==0.10.0
aif360==0.5.0
shap==0.43.0

# Microservices and Async
celery==5.3.4
redis==5.0.1
fastapi-limiter==0.1.6
httpx==0.25.2

# Additional utilities
python-multipart==0.0.6
pydantic-settings==2.1.0
python-dotenv==1.0.0
